Fixing errors, cleaning the code, restructuring the file tree
Fixing the error of adding only 2 columns to the dataFrame now we are able to add as many cols to dataFrame as we want. After the funcionality for the cols was added there was no need for OtherProjects() function. The App is getting bigger so we decided to restructure the file tree now you can find the classes in the Classes folder and the last change is made to the invest.py file where user can add it's input.